insacloud/requirements.txt
Django==1.8.3
django-admin-bootstrapped==2.5.4
django-bootstrap3==6.2.2
django-filter==0.11.0
django-redis-cache==1.6.0
djangorestframework==3.2.4
gunicorn==19.3.0
Markdown==2.6.2
Pillow==3.0.0
psycopg2==2.6.1
redis==2.10.3
requests==2.8.1
setproctitle==1.1.9
wheel==0.24.0